+++ loan/src/main/java/loan/easyloan/entity/RepayMoneyFlow.java
@@ -3,8 +3,8 @@ package loan.easyLoan.entity;
 
 public class RepayMoneyFlow {
 
-  private long serialNumber;
-  private long billId;
+  private int serialNumber;
+  private int billId;
   private String inBoundAccount;
   private String outBoundAccount;
   private java.sql.Timestamp exactDate;
@@ -14,20 +14,20 @@ public class RepayMoneyFlow {
   private double principalMoney;
 
 
-  public long getSerialNumber() {
+  public int getSerialNumber() {
     return serialNumber;
   }
 
-  public void setSerialNumber(long serialNumber) {
+  public void setSerialNumber(int serialNumber) {
     this.serialNumber = serialNumber;
   }
 
 
-  public long getBillId() {
+  public int getBillId() {
     return billId;
   }
 
-  public void setBillId(long billId) {
+  public void setBillId(int billId) {
     this.billId = billId;
   }
